backend/requirements.txt
uvicorn
pika
fastapi
mysql-connector-python
email-validator